+++ src/service/EpicHandler.java
@@ -12,6 +12,7 @@ import exceptions.ManagerNotAcceptableException;
 import exceptions.ManagerNotFoundException;
 import exceptions.ManagerSaveException;
 import model.Epic;
+import service.handlerbase.BaseHttpHandler;
 
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
