+++ src/service/TaskHandler.java
@@ -12,6 +12,7 @@ import exceptions.ManagerNotAcceptableException;
 import exceptions.ManagerNotFoundException;
 import exceptions.ManagerSaveException;
 import model.Task;
+import service.handlerbase.BaseHttpHandler;
 
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
